app/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1
boto3==1.38.31
pyjwt==2.12.1
pwdlib[argon2]==0.9.0